chore: move format functions to a separate file
