+++ src/duration.rs
@@ -73,10 +73,6 @@ impl fmt::Display for Error {
     }
 }
 
-/// A wrapper type that allows you to Display a Duration
-#[derive(Debug, Clone)]
-pub struct FormattedDuration(Duration);
-
 trait OverflowOp: Sized {
     fn mul(self, other: Self) -> Result<Self, Error>;
     fn add(self, other: Self) -> Result<Self, Error>;
@@ -234,107 +230,16 @@ pub fn parse_duration(s: &str) -> Result<Duration, Error> {
     .parse()
 }
 
-/// Formats duration into a human-readable string
-///
-/// Note: this format is guaranteed to have same value when using
-/// parse_duration, but we can change some details of the exact composition
-/// of the value.
-///
-/// # Examples
-///
-/// ```
-/// use std::time::Duration;
-/// use humantime::format_duration;
-///
-/// let val1 = Duration::new(9420, 0);
-/// assert_eq!(format_duration(val1).to_string(), "2h 37m");
-/// let val2 = Duration::new(0, 32_000_000);
-/// assert_eq!(format_duration(val2).to_string(), "32ms");
-/// ```
-pub fn format_duration(val: Duration) -> FormattedDuration {
-    FormattedDuration(val)
-}
-
-fn item_plural(f: &mut fmt::Formatter, started: &mut bool, name: &str, value: u64) -> fmt::Result {
-    if value > 0 {
-        if *started {
-            f.write_str(" ")?;
-        }
-        write!(f, "{}{}", value, name)?;
-        if value > 1 {
-            f.write_str("s")?;
-        }
-        *started = true;
-    }
-    Ok(())
-}
-fn item(f: &mut fmt::Formatter, started: &mut bool, name: &str, value: u32) -> fmt::Result {
-    if value > 0 {
-        if *started {
-            f.write_str(" ")?;
-        }
-        write!(f, "{}{}", value, name)?;
-        *started = true;
-    }
-    Ok(())
-}
-
-impl FormattedDuration {
-    /// Returns a reference to the [`Duration`][] that is being formatted.
-    pub fn get_ref(&self) -> &Duration {
-        &self.0
-    }
-}
-
-impl fmt::Display for FormattedDuration {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let secs = self.0.as_secs();
-        let nanos = self.0.subsec_nanos();
-
-        if secs == 0 && nanos == 0 {
-            f.write_str("0s")?;
-            return Ok(());
-        }
-
-        let years = secs / 31_557_600; // 365.25d
-        let ydays = secs % 31_557_600;
-        let months = ydays / 2_630_016; // 30.44d
-        let mdays = ydays % 2_630_016;
-        let days = mdays / 86400;
-        let day_secs = mdays % 86400;
-        let hours = day_secs / 3600;
-        let minutes = day_secs % 3600 / 60;
-        let seconds = day_secs % 60;
-
-        let millis = nanos / 1_000_000;
-        let micros = nanos / 1000 % 1000;
-        let nanosec = nanos % 1000;
-
-        let started = &mut false;
-        item_plural(f, started, "year", years)?;
-        item_plural(f, started, "month", months)?;
-        item_plural(f, started, "day", days)?;
-        item(f, started, "h", hours as u32)?;
-        item(f, started, "m", minutes as u32)?;
-        item(f, started, "s", seconds as u32)?;
-        item(f, started, "ms", millis)?;
-        #[cfg(feature = "mu")]
-        item(f, started, "µs", micros)?;
-        #[cfg(not(feature = "mu"))]
-        item(f, started, "us", micros)?;
-        item(f, started, "ns", nanosec)?;
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 mod test {
     use std::time::Duration;
 
     use rand::Rng;
 
+    use crate::format_duration;
+
+    use super::parse_duration;
     use super::Error;
-    use super::{format_duration, parse_duration};
 
     #[test]
     #[allow(clippy::cognitive_complexity)]
