+++ src/lib.rs
@@ -20,6 +20,7 @@
 
 mod date;
 mod duration;
+pub mod format;
 pub mod new;
 mod wrapper;
 
@@ -29,6 +30,6 @@ pub use self::date::{
     format_rfc3339_seconds,
 };
 pub use self::date::{parse_rfc3339, parse_rfc3339_weak, Error as TimestampError};
-pub use self::duration::{format_duration, FormattedDuration};
 pub use self::duration::{parse_duration, Error as DurationError};
+pub use self::format::{format_duration, FormattedDuration};
 pub use self::wrapper::{Duration, Timestamp};
